Retrieval/requirements.txt
faiss-cpu
fastapi
open_clip_torch[training]
uvicorn
openpyxl
pandas